Safari-609.1.11.1/LayoutTests/http/tests/security/cross-frame-access-protocol-expected.txt

CONSOLE MESSAGE: line 194: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.

PASS: Cross frame access to https from http was denied!


--------
Frame: 'aFrame'
--------
Inner iframe.
